Update csproj; add stylecop and comment docs
