+++ .editorconfig
@@ -1,5 +1,6 @@
 root = true
 
+# dotnet-format Rules
 [*.cs]
 indent_style = space
 indent_size = 4
@@ -8,3 +9,30 @@ insert_final_newline = true
 trim_trailing_whitespace = true
 dotnet_sort_system_directives_first = true
 dotnet_separate_import_directive_groups = true
+max_line_length = 120
+
+# StyleCop Rules
+# Don't require file headers
+dotnet_diagnostic.SA1633.severity = none
+# Using directives placement
+dotnet_diagnostic.SA1200.severity = none
+# Prefix local calls with "this."
+dotnet_diagnostic.SA1101.severity = none
+# Field names should not begin with underscore
+dotnet_diagnostic.SA1309.severity = none
+
+# Require public members to be documented
+dotnet_diagnostic.SA1600.severity = warning
+# Disable docs requirement for internal and private members
+dotnet_diagnostic.SA1600.disabled = true
+
+# Curly brackets for if/while/etc
+dotnet_diagnostic.SA1503.severity = error
+# Elements should be separated by blank line
+dotnet_diagnostic.SA1516.severity = error
+# Classes should have explicit access modifiers
+dotnet_diagnostic.SA1400.severity = error
+# Fields should have explicit access modifiers
+dotnet_diagnostic.SA1401.severity = error
+# Enforce a max line length of 120 characters in StyleCop
+dotnet_diagnostic.SA1027.severity = error
