+++ src/Wristband.AspNet.Auth.csproj
@@ -9,6 +9,10 @@
     <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" Condition="'$(TargetFramework)' == 'net6.0'" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" Condition="'$(TargetFramework)' == 'net7.0'" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" Condition="'$(TargetFramework)' == 'net8.0'" />
+    <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.556">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
   </ItemGroup>
 
   <ItemGroup>
@@ -17,6 +21,7 @@
   </ItemGroup>
 
   <PropertyGroup>
+    <PackageId>Wristband.AspNet.Auth</PackageId>
     <Product>Wristband.AspNet.Auth</Product>
     <Description>SDK for integrating your ASP.NET web application with Wristband. Handles user authentication and token management.</Description>
     <Copyright>Apitopia, Inc.</Copyright>
@@ -28,11 +33,13 @@
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/wristband-dev/aspnet-auth</RepositoryUrl>
+    <RepositoryBranch>main</RepositoryBranch>
     <CLSCompliant>true</CLSCompliant>
     <ComVisible>false</ComVisible>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <PackageReadmeFile>README.md</PackageReadmeFile>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup>
